requirements.txt
jupyter
pandas
tqdm
tabulate
matplotlib
fire==0.2.1

pyyaml
seaborn

sentence-transformers==0.2.6.1

nltk==3.5
node2vec==0.3.2
numpy==1.18.3
networkx==2.4

torch==1.2.0
tokenizers==0.7.0
transformers==2.9.1
karateclub==1.0.0
gensim==3.8.2
scikit-learn==0.22.2.post1

git+https://github.com/allenai/longformer.git
